dropped QOpenGLTexture for raw textures

+++ rendering/framebufferobject.cpp
@@ -66,7 +66,7 @@ void FramebufferObject::Create(QOpenGLContext *ctx, int width, int height)
         );*/
 
   ctx->functions()->glTexImage2D(
-        GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, GL_RGBA,  GL_FLOAT, nullptr
+        GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA,  GL_FLOAT, nullptr
         );
 
   // set texture filtering to bilinear
